Split handler.go to multiple files
