+++ pkg/model/call_request.go
@@ -1,0 +1,8 @@
+package model
+
+//CallRequest godoc
+type CallRequest struct {
+	URL    string      `json:"url"`
+	Method string      `json:"method"`
+	Body   interface{} `json:"body"`
+}
